fix: scope inclusion and include first label in scope name
Scope inclusion was computed by checking prefixes of the full name.
This could cause wrong inclusion when the label was a prefix of
another first label, for example with `test` and `testing`.
Inclusion is now computed by checking prefixes of the list of labels.

Scope name always skipped the first label, this is now fixed.